+++ docs/pages/start.js
@@ -225,13 +225,13 @@ class Start extends Component {
                     <div className="start-box-features-inner">
                         <Flex>
                             <FlexItem>
-                                <Link to="/docs/2/articles/0" className="start-box-features-item color-green">
+                                <Link to="/page/2/articles/0" className="start-box-features-item color-green">
                                     <FontAwesome name="play-circle" size="4x"/>
                                     <p>{ lang.getstart }</p>
                                 </Link>
                             </FlexItem>
                             <FlexItem>
-                                <Link to="/docs/1/articles/0" className="start-box-features-item color-blue">
+                                <Link to="/page/1/articles/0" className="start-box-features-item color-blue">
                                     <FontAwesome name="th" size="4x" />
                                     <p>{ lang.component }</p>
                                 </Link>
@@ -249,7 +249,7 @@ class Start extends Component {
                     <div className="start-box-footer-inner">
                         <Footer>
                             <FooterText>{ lang.footer }</FooterText>
-                            <FooterText>Copyright &copy; 2008-2016 weui.io</FooterText>
+                            <FooterText>Copyright &copy; 2008-2017 weui.io</FooterText>
                         </Footer>
                     </div>
                 </div>