+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Breaking Changes
 - Remove mobile_detect utilities, ActionSheet and Dialog will needs to provide a platform by user.
 - Used standard prop-types library
+- css file now in `react-weui/build/packages/react-weui.css`
 
 Changes